testing/requirements.txt
# Testing Framework Dependencies
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-cov==4.1.0
pytest-mock==3.12.0
webdriver-manager==4.0.1
allure-pytest==2.13.2
faker==20.1.0
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.1.4
numpy==1.25.2
Pillow==10.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
